added mapbox map to project. Added new list and map view tabs for search
