+++ package.json
@@ -11,11 +11,13 @@
   },
   "dependencies": {
     "bootstrap": "^5.3.3",
+    "mapbox-gl": "^3.10.0",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.9",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
     "react-router-dom": "^7.1.2",
+    "react-tabs": "^6.1.0",
     "serve": "^14.2.4"
   },
   "devDependencies": {
